# ext/bcmath: Improving `bcpow()` performance