--- a/ext/bcmath/libbcmath/src/raise.c
+++ b/ext/bcmath/libbcmath/src/raise.c
@@ -199,12 +199,12 @@ bc_raise_status bc_raise(bc_num base, long exponent, bc_num *result, size_t scal
 	}
 
 	/* check overflow */
-	if (UNEXPECTED(base->n_len > SIZE_T_MAX / exponent)) {
+	if (UNEXPECTED(base->n_len > SIZE_MAX / exponent)) {
 		bc_free_num (result);
 		*result = bc_copy_num(BCG(_one_));
 		return BC_RAISE_STATUS_LEN_IS_OVERFLOW;
 	}
-	if (UNEXPECTED(base->n_scale > SIZE_T_MAX / exponent)) {
+	if (UNEXPECTED(base->n_scale > SIZE_MAX / exponent)) {
 		bc_free_num (result);
 		*result = bc_copy_num(BCG(_one_));
 		return BC_RAISE_STATUS_SCALE_IS_OVERFLOW;
@@ -215,7 +215,7 @@ bc_raise_status bc_raise(bc_num base, long exponent, bc_num *result, size_t scal
 	size_t power_scale = base->n_scale * exponent;
 
 	/* check overflow */
-	if (UNEXPECTED(power_len > SIZE_T_MAX - power_scale)) {
+	if (UNEXPECTED(power_len > SIZE_MAX - power_scale)) {
 		bc_free_num (result);
 		*result = bc_copy_num(BCG(_one_));
 		return BC_RAISE_STATUS_FULLLEN_IS_OVERFLOW;