--- a/ext/bcmath/libbcmath/src/recmul.c
+++ b/ext/bcmath/libbcmath/src/recmul.c
@@ -72,28 +72,6 @@ static inline void bc_fast_mul(bc_num n1, size_t n1len, bc_num n2, size_t n2len,
 	}
 }
 
-/*
- * Equivalent of bc_fast_mul for small numbers to perform computations
- * without using array.
- */
-static inline void bc_fast_square(bc_num n1, size_t n1len, bc_num *prod)
-{
-	const char *n1end = n1->n_value + n1len - 1;
-
-	BC_VECTOR n1_vector = bc_partial_convert_to_vector(n1end, n1len);
-	BC_VECTOR prod_vector = n1_vector * n1_vector;
-
-	size_t prodlen = n1len + n1len;
-	*prod = bc_new_num_nonzeroed(prodlen, 0);
-	char *pptr = (*prod)->n_value;
-	char *pend = pptr + prodlen - 1;
-
-	while (pend >= pptr) {
-		*pend-- = prod_vector % BASE;
-		prod_vector /= BASE;
-	}
-}
-
 /* Common part of functions bc_standard_mul and bc_standard_square
  * that takes a vector and converts it to a bc_num 	*/
 static inline void bc_mul_finish_from_vector(BC_VECTOR *prod_vector, size_t prod_arr_size, size_t prodlen, bc_num *prod) {
@@ -179,55 +157,6 @@ static void bc_standard_mul(bc_num n1, size_t n1len, bc_num n2, size_t n2len, bc
 		efree(n1_vector);
 	}
 }
-
-/** This is bc_standard_mul implementation for square */
-static void bc_standard_square(bc_num n1, size_t n1len, bc_num *prod)
-{
-	size_t i;
-	const char *n1end = n1->n_value + n1len - 1;
-	size_t prodlen = n1len + n1len;
-
-	size_t n1_arr_size = BC_ARR_SIZE_FROM_LEN(n1len);
-	size_t prod_arr_size = BC_ARR_SIZE_FROM_LEN(prodlen);
-
-	BC_VECTOR *buf = safe_emalloc(n1_arr_size + n1_arr_size + prod_arr_size, sizeof(BC_VECTOR), 0);
-
-	BC_VECTOR *n1_vector = buf;
-	BC_VECTOR *prod_vector = n1_vector + n1_arr_size + n1_arr_size;
-
-	for (i = 0; i < prod_arr_size; i++) {
-		prod_vector[i] = 0;
-	}
-
-	/* Convert to BC_VECTOR[] */
-	bc_convert_to_vector(n1_vector, n1end, n1len);
-
-	/* Multiplication and addition */
-	size_t count = 0;
-	for (i = 0; i < n1_arr_size; i++) {
-		/*
-		 * This calculation adds the result multiple times to the array entries.
-		 * When multiplying large numbers of digits, there is a possibility of
-		 * overflow, so digit adjustment is performed beforehand.
-		 */
-		if (UNEXPECTED(count >= BC_VECTOR_NO_OVERFLOW_ADD_COUNT)) {
-			bc_mul_carry_calc(prod_vector, prod_arr_size);
-			count = 0;
-		}
-		count++;
-		for (size_t j = 0; j < n1_arr_size; j++) {
-			prod_vector[i + j] += n1_vector[i] * n1_vector[j];
-		}
-	}
-
-	bc_mul_finish_from_vector(prod_vector, prod_arr_size, prodlen, prod);
-
-	efree(buf);
-}
-
-/* The multiply routine. N2 times N1 is put int PROD with the scale of
-   the result being MIN(N2 scale+N1 scale, MAX (SCALE, N2 scale, N1 scale)).
-   */
 
 bc_num bc_multiply(bc_num n1, bc_num n2, size_t scale)
 {